feat(dotcom): redesign admin panel and add file undelete (#9792)
In order to make the admin panel usable as it grows (it was one long
crowded page) and to let us restore files users deleted by accident — a
recurring support ask — this PR splits the panel into sidebar-navigated
sections, adds a soft-deleted-files view with undelete, and reskins the
page in the intel design language. Closes #7571.

The work is deliberately staged as independent commits so the visual
redesign can be reverted alone:

1. `refactor`: split into `/admin/:section?` subroutes (Users, Files &
templates, Feature flags, System) with a left sidebar; pure
rearrangement of the existing components.
2. `feat`: undelete. Two review-relevant subtleties: a user's own
deleted files never reach their replicated store (`fetchEverythingSql`
filters `isDeleted = false`), so the deleted-files table reads from a
dedicated pg-backed admin endpoint scoped to files the user owns (legacy
`ownerId`, home workspace, or `owner`-role workspaces). And restoring
flips `isDeleted` + recreates `file_state`/`group_file`, then
hard-reboots every affected user DO — without the reboot, live sessions
never see the restored file because the replicator has no subscription
edge for it at fan-out time.
3. `feat`: intel reskin, confined to `pages/admin/` (tokens ported into
the CSS module, local `AdminButton`); `git revert` of this commit
restores the TLA look with everything else intact.

Both new routes sit behind the existing `/app/admin/*` middleware
(verified `@tldraw.com` email); the UI keeps the `user.isTldraw` gate.

### Change type

- [x] `feature`

### Test plan

1. Visit `/admin` as a tldraw user — redirects to `/admin/users`;
sections deep-link and survive refresh.
2. Soft-delete one of your files, look yourself up in Users — the file
appears under "Deleted files" with workspace/role; undelete it and
confirm it reappears in your sidebar without a reload.
3. Files & templates → "Undelete file" restores a known file ID
directly; unknown ID → 404, live file → 400.
4. Revert the reskin commit locally and confirm the panel still works in
the old TLA styling.

- [x] Unit tests

### Code changes

| Section | LOC change |
| ------- | ---------- |
| Apps    | +1933 / -1443 |
| Tests   | +194 / -0  |
