--- apps/dotcom/client/src/pages/admin.module.css
@@ -1,414 +1,0 @@
-.adminContainer {
-	display: flex;
-	flex-direction: column;
-	height: 100vh;
-	overflow: auto;
-	background-color: var(--tla-color-sidebar);
-	color: var(--tl-color-text-0);
-	font-family: var(--tla-font-ui);
-	max-width: 100vw;
-}
-
-.adminHeader {
-	padding: 24px 32px 16px 32px;
-	border-bottom: 1px solid var(--tl-color-divider);
-	background-color: var(--tl-color-panel);
-}
-
-.adminReleaseMeta {
-	margin: 10px 0 0 0;
-	font-size: 12px;
-	line-height: 1.4;
-	color: var(--tl-color-text-3);
-}
-
-.adminReleaseLabel {
-	font-weight: 500;
-	margin-right: 4px;
-}
-
-.adminReleaseValue {
-	font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, monospace;
-	font-size: 11px;
-	font-weight: 500;
-	color: var(--tl-color-text-0);
-	word-break: break-all;
-}
-
-.adminContent {
-	flex: 1;
-	padding: 24px 32px;
-	display: flex;
-	flex-direction: column;
-	gap: var(--tl-space-8);
-	overflow: auto;
-}
-
-.adminSection {
-	display: flex;
-	flex-direction: column;
-	gap: var(--tl-space-5);
-}
-
-.searchContainer {
-	display: flex;
-	gap: var(--tl-space-4);
-	align-items: center;
-}
-
-.searchInput {
-	height: 32px;
-	padding: 0 12px;
-	border: 1px solid var(--tl-color-divider);
-	border-radius: var(--tl-radius-1);
-	background-color: var(--tl-color-selected-contrast);
-	color: var(--tl-color-text-0);
-	font-family: var(--tla-font-ui);
-	font-size: 12px;
-	font-weight: 500;
-	outline: none;
-}
-
-.searchInput:focus {
-	border-color: var(--tl-color-primary);
-}
-
-.searchInput:disabled {
-	opacity: 0.4;
-	cursor: not-allowed;
-}
-
-.errorMessage {
-	color: var(--tl-color-danger);
-	font-size: 12px;
-	font-weight: 500;
-	padding: 8px 12px;
-	background-color: var(--tla-color-secondary);
-	border-radius: var(--tl-radius-1);
-	border: 1px solid var(--tl-color-danger);
-}
-
-.successMessage {
-	color: var(--tl-color-success);
-	font-size: 12px;
-	font-weight: 500;
-	padding: 8px 12px;
-	background-color: var(--tla-color-secondary);
-	border-radius: var(--tl-radius-1);
-	border: 1px solid var(--tl-color-success);
-}
-
-.userActions {
-	display: flex;
-	gap: var(--tl-space-4);
-	align-items: center;
-}
-
-.userActionButton {
-	border: 1px solid transparent !important;
-}
-
-.userActionButton:hover:not(:disabled) {
-	border-color: var(--tl-color-text-3) !important;
-}
-
-.dataDisplay {
-	background-color: var(--tl-color-low);
-	border: 1px solid var(--tl-color-divider);
-	border-radius: var(--tl-radius-1);
-	padding: var(--tl-space-5);
-	font-family: 'IBMPlexMono', monospace;
-	font-size: 11px;
-	line-height: 1.4;
-	color: var(--tl-color-text-0);
-	overflow: auto;
-	max-height: 400px;
-	user-select: text;
-}
-
-.fileOperations {
-	display: flex;
-	flex-direction: column;
-	gap: var(--tl-space-5);
-}
-
-.fileOperation {
-	display: flex;
-	flex-direction: column;
-	gap: var(--tl-space-4);
-	padding: var(--tl-space-5);
-	background-color: var(--tl-color-panel);
-	border: 1px solid var(--tl-color-divider);
-	border-radius: var(--tl-radius-1);
-}
-
-.downloadContainer {
-	display: flex;
-	gap: var(--tl-space-4);
-	align-items: center;
-}
-
-.dangerZone {
-	display: flex;
-	flex-direction: column;
-	gap: var(--tl-space-4);
-	padding: var(--tl-space-5);
-	background-color: var(--tl-color-panel);
-	border: 1px solid var(--tl-color-danger);
-	border-radius: var(--tl-radius-1);
-}
-
-.deleteContainer {
-	display: flex;
-	gap: var(--tl-space-4);
-	align-items: center;
-}
-
-.deleteButton {
-	border: 2px solid var(--tl-color-danger) !important;
-}
-
-/* Progress Log */
-.progressLog {
-	margin-top: 16px;
-	padding: var(--tl-space-5);
-	background-color: var(--tl-color-low);
-	border: 1px solid var(--tl-color-divider);
-	border-radius: var(--tl-radius-1);
-}
-
-.progressLog h5 {
-	margin: 0 0 12px 0;
-	font-size: 12px;
-	font-weight: 600;
-	color: var(--tl-color-text-0);
-	text-transform: uppercase;
-	letter-spacing: 0.5px;
-}
-
-.logContainer {
-	max-height: 200px;
-	overflow-y: auto;
-	font-family: 'IBMPlexMono', monospace;
-	font-size: 11px;
-	line-height: 1.4;
-	color: var(--tl-color-text-0);
-	background-color: var(--tl-color-selected-contrast);
-	border: 1px solid var(--tl-color-divider);
-	border-radius: var(--tl-radius-1);
-	padding: var(--tl-space-3);
-}
-
-.logEntry {
-	margin-bottom: 4px;
-	padding: 2px 0;
-	border-bottom: 1px solid var(--tl-color-divider);
-}
-
-.logEntry:last-child {
-	border-bottom: none;
-	margin-bottom: 0;
-}
-
-/* Structured Data Display */
-.structuredData {
-	position: relative;
-	display: flex;
-	flex-direction: column;
-	gap: var(--tl-space-4);
-}
-
-.copyButton {
-	position: absolute;
-	top: 8px;
-	right: 8px;
-	z-index: 10;
-}
-
-/* User Summary */
-.userSummary {
-	padding: var(--tl-space-5);
-	background-color: var(--tl-color-panel);
-	border: 1px solid var(--tl-color-divider);
-	border-radius: var(--tl-radius-1);
-}
-
-.summaryGrid {
-	display: grid;
-	grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-	gap: var(--tl-space-4);
-}
-
-.summaryItem {
-	display: flex;
-	flex-direction: column;
-	gap: var(--tl-space-2);
-}
-
-.fieldLabel {
-	font-size: 11px;
-	font-weight: 600;
-	color: var(--tl-color-text-3);
-	text-transform: uppercase;
-	letter-spacing: 0.5px;
-}
-
-.fieldValue {
-	font-size: 12px;
-	font-weight: 500;
-	color: var(--tl-color-text-0);
-	word-break: break-word;
-}
-
-/* Migration Section */
-.migrationSection {
-	display: flex;
-	flex-direction: column;
-	gap: var(--tl-space-4);
-	padding: var(--tl-space-5);
-	background-color: var(--tl-color-panel);
-	border: 1px solid var(--tl-color-primary);
-	border-radius: var(--tl-radius-1);
-	margin-top: 16px;
-}
-
-/* Stats Display */
-.statsContainer {
-	display: flex;
-	gap: var(--tl-space-5);
-	padding: var(--tl-space-4);
-	flex-wrap: wrap;
-}
-
-.statItem {
-	display: flex;
-	flex-direction: column;
-	gap: var(--tl-space-2);
-}
-
-.statLabel {
-	font-size: 11px;
-	font-weight: 600;
-	color: var(--tl-color-text-3);
-	text-transform: uppercase;
-}
-
-.statValue {
-	font-size: 16px;
-	font-weight: 600;
-	color: var(--tl-color-text-0);
-}
-
-/* Count Container */
-.countContainer {
-	display: flex;
-	gap: var(--tl-space-4);
-	align-items: center;
-}
-
-.countDisplay {
-	font-size: 12px;
-	color: var(--tl-color-text-0);
-}
-
-/* Config Container */
-.configContainer {
-	display: flex;
-	gap: var(--tl-space-5);
-	flex-wrap: wrap;
-}
-
-/* Feature Flags */
-.featureFlagsNote {
-	margin-bottom: 8px;
-}
-
-.featureFlagsDescription {
-	color: var(--tl-color-text-3);
-}
-
-.featureFlagsContainer {
-	display: flex;
-	flex-direction: column;
-	gap: var(--tl-space-5);
-	margin-top: 16px;
-}
-
-.featureFlagItem {
-	display: flex;
-	align-items: center;
-	gap: var(--tl-space-4);
-}
-
-.featureFlagLabel {
-	display: flex;
-	align-items: center;
-	gap: var(--tl-space-3);
-	cursor: pointer;
-	min-width: 200px;
-}
-
-.featureFlagLabel button:disabled {
-	opacity: 0.4;
-	cursor: not-allowed;
-}
-
-.featureFlagDisabled {
-	opacity: 0.4;
-	cursor: not-allowed;
-}
-
-/* Batch Migration */
-.sleepInput {
-	width: 100px;
-	margin-left: 8px;
-}
-
-/* Responsive adjustments */
-@media (max-width: 768px) {
-	.adminHeader {
-		padding: 16px 20px 12px 20px;
-	}
-
-	.adminContent {
-		padding: 16px 20px;
-		gap: 24px;
-	}
-
-	.searchContainer,
-	.downloadContainer,
-	.deleteContainer {
-		flex-direction: column;
-		align-items: stretch;
-	}
-
-	.userActions {
-		flex-direction: column;
-		align-items: stretch;
-	}
-
-	.summaryGrid {
-		grid-template-columns: 1fr;
-	}
-}
-
-/* Asset diagnostics */
-.diagnosticsTable {
-	width: 100%;
-	border-collapse: collapse;
-	font-size: 12px;
-	font-family: monospace;
-}
-
-.diagnosticsTable th,
-.diagnosticsTable td {
-	text-align: left;
-	padding: var(--tl-space-2) var(--tl-space-3);
-	border-bottom: 1px solid var(--tl-color-divider);
-	word-break: break-all;
-}
-
-.diagnosticsTable th {
-	font-weight: 600;
-	color: var(--tl-color-text-3);
-}
