--- apps/dotcom/client/src/pages/admin.tsx
@@ -1,1028 +1,0 @@
-import {
-	AdminFileAssetsResponseBody,
-	FeatureFlagValue,
-	PercentageFeatureFlag,
-	TlaFile,
-	ZStoreData,
-} from '@tldraw/dotcom-shared'
-import { useCallback, useEffect, useRef, useState } from 'react'
-import { Navigate } from 'react-router-dom'
-import { fetch } from 'tldraw'
-import { sentryReleaseName } from '../../sentry-release-name'
-import { TlaButton } from '../tla/components/TlaButton/TlaButton'
-import { useTldrawCurrentUser } from '../tla/hooks/useUser'
-import styles from './admin.module.css'
-
-// Helper component for structured data display.
-function StructuredDataDisplay({ data }: { data: object }) {
-	const [copied, setCopied] = useState(false)
-
-	const handleCopy = async () => {
-		try {
-			await navigator.clipboard.writeText(JSON.stringify(data, null, 2))
-			setCopied(true)
-			setTimeout(() => setCopied(false), 2000)
-		} catch (err) {
-			console.error('Failed to copy:', err)
-		}
-	}
-
-	return (
-		<div className={styles.structuredData}>
-			<TlaButton onClick={handleCopy} variant="secondary" className={styles.copyButton}>
-				{copied ? 'Copied!' : 'Copy JSON'}
-			</TlaButton>
-			<pre className={styles.dataDisplay}>{JSON.stringify(data, null, 2)}</pre>
-		</div>
-	)
-}
-
-// Helper component for user data summary
-function UserDataSummary({ data }: { data: ZStoreData }) {
-	const getUserInfo = () => {
-		const user = data.user[0]
-		const files = data.file || []
-		const deletedFiles = files.filter((f: TlaFile) => f.isDeleted)
-		const activeFiles = files.filter((f: TlaFile) => !f.isDeleted)
-
-		return {
-			name: user?.name || 'Unknown',
-			email: user?.email || 'No email',
-			activeFiles: activeFiles.length,
-			deletedFiles: deletedFiles.length,
-		}
-	}
-
-	const info = getUserInfo()
-
-	return (
-		<div className={styles.userSummary}>
-			<div className={styles.summaryGrid}>
-				<div className={styles.summaryItem}>
-					<span className={styles.fieldLabel}>Name:</span>
-					<span className={styles.fieldValue}>{info.name}</span>
-				</div>
-				<div className={styles.summaryItem}>
-					<span className={styles.fieldLabel}>Email:</span>
-					<span className={styles.fieldValue}>{info.email}</span>
-				</div>
-				<div className={styles.summaryItem}>
-					<span className={styles.fieldLabel}>Active Files:</span>
-					<span className={styles.fieldValue}>{info.activeFiles}</span>
-				</div>
-				<div className={styles.summaryItem}>
-					<span className={styles.fieldLabel}>Deleted Files:</span>
-					<span className={styles.fieldValue}>{info.deletedFiles}</span>
-				</div>
-			</div>
-		</div>
-	)
-}
-
-export function Component() {
-	const user = useTldrawCurrentUser()
-	const [data, setData] = useState<any>(null)
-	const [error, setError] = useState(null as string | null)
-	const [replicatorData, setReplicatorData] = useState(null)
-	const [isRebooting, setIsRebooting] = useState(false)
-	const [successMessage, setSuccessMessage] = useState(null as string | null)
-	const inputRef = useRef<HTMLInputElement>(null)
-
-	const loadData = useCallback(async () => {
-		const q = inputRef.current?.value?.trim() ?? ''
-		if (!q) {
-			setError('Please enter an email or ID')
-			return
-		}
-
-		setError(null)
-		setSuccessMessage(null)
-
-		const res = await fetch(`/api/app/admin/user?${new URLSearchParams({ q })}`)
-		if (!res.ok) {
-			setError(res.statusText + ': ' + (await res.text()))
-			return
-		}
-		setError(null)
-		setData(await res.json())
-	}, [])
-
-	const doReboot = useCallback(async () => {
-		const q = inputRef.current?.value?.trim() ?? ''
-		if (!q) {
-			setError('Please enter an email or ID')
-			return
-		}
-
-		setIsRebooting(true)
-		setError(null)
-		try {
-			const res = await fetch(`/api/app/admin/user/reboot?${new URLSearchParams({ q })}`, {
-				method: 'POST',
-			})
-			if (!res.ok) {
-				setError(res.statusText + ': ' + (await res.text()))
-				return
-			}
-			setError(null)
-			loadData()
-			setSuccessMessage('User rebooted successfully')
-		} finally {
-			setIsRebooting(false)
-		}
-	}, [loadData])
-
-	useEffect(() => {
-		if (user?.isTldraw) {
-			fetch('/api/app/admin/replicator')
-				.then(async (res) => {
-					if (!res.ok) {
-						setError(res.statusText + ': ' + (await res.text()))
-						return
-					}
-					setError(null)
-					setReplicatorData(await res.json())
-				})
-				.catch((e) => {
-					setError(e.message)
-				})
-		}
-	}, [user?.isTldraw])
-
-	// Clear success message after 3 seconds
-	useEffect(() => {
-		if (successMessage) {
-			const timer = setTimeout(() => setSuccessMessage(null), 3000)
-			return () => clearTimeout(timer)
-		}
-	}, [successMessage])
-
-	if (!user?.isTldraw) {
-		return <Navigate to="/" replace />
-	}
-
-	return (
-		<div className={styles.adminContainer}>
-			<header className={styles.adminHeader}>
-				<h1 className="tla-text_ui__big">Admin Panel</h1>
-				<p className={styles.adminReleaseMeta}>
-					<span className={styles.adminReleaseLabel}>Release:</span>{' '}
-					<code className={styles.adminReleaseValue} translate="no">
-						{sentryReleaseName}
-					</code>
-				</p>
-			</header>
-
-			<main className={styles.adminContent}>
-				{/* User Search Section */}
-				<section className={styles.adminSection}>
-					<h2 className="tla-text_ui__title">User Management</h2>
-					<div className={styles.searchContainer}>
-						<input
-							ref={inputRef}
-							type="text"
-							placeholder="Email or ID"
-							onKeyDown={(e) => {
-								if (e.key === 'Enter') {
-									loadData()
-								}
-							}}
-							className={styles.searchInput}
-						/>
-						<TlaButton onClick={loadData} variant="primary">
-							Find User
-						</TlaButton>
-					</div>
-					{error && <div className={styles.errorMessage}>{error}</div>}
-					{successMessage && <div className={styles.successMessage}>{successMessage}</div>}
-				</section>
-
-				{/* User Data Section */}
-				{data && (
-					<section className={styles.adminSection}>
-						<h3 className="tla-text_ui__title">User Data</h3>
-						<UserDataSummary data={data} />
-						<div className={styles.userActions}>
-							<TlaButton
-								onClick={() => {
-									navigator.clipboard.writeText(JSON.stringify(data, null, 2))
-									setSuccessMessage('User data copied to clipboard')
-								}}
-								variant="secondary"
-							>
-								Copy Data
-							</TlaButton>
-							<TlaButton
-								disabled={isRebooting}
-								onClick={doReboot}
-								isLoading={isRebooting}
-								className={styles.userActionButton}
-							>
-								Force Reboot
-							</TlaButton>
-						</div>
-						<StructuredDataDisplay data={data} />
-					</section>
-				)}
-
-				{/* System Data Section */}
-				{replicatorData && (
-					<section className={styles.adminSection}>
-						<h3 className="tla-text_ui__title">System Health</h3>
-						<StructuredDataDisplay data={replicatorData} />
-					</section>
-				)}
-
-				{/* Feature Flags Section */}
-				<section className={styles.adminSection}>
-					<h3 className="tla-text_ui__title">Feature Flags</h3>
-					<FeatureFlags />
-				</section>
-
-				{/* Welcome Template Section */}
-				<section className={styles.adminSection}>
-					<h3 className="tla-text_ui__title">Welcome template</h3>
-					<WelcomeTemplate />
-				</section>
-
-				{/* File Operations Section */}
-				<section className={styles.adminSection}>
-					<h3 className="tla-text_ui__title">File Operations</h3>
-					<div className={styles.fileOperations}>
-						<DownloadTldrFile legacy={false} />
-						<DownloadTldrFile legacy={true} />
-						<CreateLegacyFile />
-						<AssetDiagnostics />
-					</div>
-				</section>
-
-				{/* Danger Zone Section */}
-				<section className={styles.adminSection}>
-					<h3 className="tla-text_ui__title">Danger Zone</h3>
-					<HardDeleteFile />
-					<DeleteUser />
-				</section>
-			</main>
-		</div>
-	)
-}
-
-function WelcomeTemplate() {
-	const inputRef = useRef<HTMLInputElement>(null)
-	const [current, setCurrent] = useState(
-		null as { fileId: string; publishedSlug: string; live?: boolean } | null
-	)
-	const [isLoading, setIsLoading] = useState(true)
-	const [error, setError] = useState(null as string | null)
-	const [successMessage, setSuccessMessage] = useState(null as string | null)
-
-	const load = useCallback(async () => {
-		setIsLoading(true)
-		setError(null)
-		try {
-			const res = await fetch('/api/app/admin/welcome-template')
-			if (!res.ok) {
-				setError(res.statusText + ': ' + (await res.text()))
-				return
-			}
-			setCurrent(await res.json())
-		} catch (err) {
-			setError(err instanceof Error ? err.message : 'Failed to load welcome template')
-		} finally {
-			setIsLoading(false)
-		}
-	}, [])
-
-	useEffect(() => {
-		load()
-	}, [load])
-
-	const onSet = useCallback(async () => {
-		const fileId = inputRef.current?.value?.trim()
-		if (!fileId) {
-			setError('Please enter a published file ID')
-			return
-		}
-		setError(null)
-		setSuccessMessage(null)
-		try {
-			const res = await fetch('/api/app/admin/welcome-template', {
-				method: 'POST',
-				headers: { 'Content-Type': 'application/json' },
-				body: JSON.stringify({ fileId }),
-			})
-			if (!res.ok) {
-				setError(res.statusText + ': ' + (await res.text()))
-				return
-			}
-			setCurrent(await res.json())
-			setSuccessMessage('Welcome template set ✨')
-			inputRef.current!.value = ''
-		} catch (err) {
-			setError(err instanceof Error ? err.message : 'Failed to set welcome template')
-		}
-	}, [])
-
-	const onClear = useCallback(async () => {
-		if (
-			!window.confirm('Clear the welcome template? New workspaces will use the built-in default.')
-		)
-			return
-		setError(null)
-		setSuccessMessage(null)
-		try {
-			const res = await fetch('/api/app/admin/welcome-template/clear', { method: 'POST' })
-			if (!res.ok) {
-				setError(res.statusText + ': ' + (await res.text()))
-				return
-			}
-			setCurrent(null)
-			setSuccessMessage('Welcome template cleared — using the built-in default')
-		} catch (err) {
-			setError(err instanceof Error ? err.message : 'Failed to clear welcome template')
-		}
-	}, [])
-
-	useEffect(() => {
-		if (successMessage) {
-			const timer = setTimeout(() => setSuccessMessage(null), 3000)
-			return () => clearTimeout(timer)
-		}
-	}, [successMessage])
-
-	return (
-		<div className={styles.fileOperation}>
-			<p className="tla-text_ui__regular">
-				The file new workspaces fork their first file from. Publish the file first, then set it here
-				by its file ID. Clear it to use the built-in default.
-			</p>
-			{error && <div className={styles.errorMessage}>{error}</div>}
-			{successMessage && <div className={styles.successMessage}>{successMessage}</div>}
-			<div className={styles.summaryItem}>
-				<span className={styles.fieldLabel}>Current:</span>
-				<span className={styles.fieldValue}>
-					{isLoading
-						? 'Loading…'
-						: current
-							? `${current.fileId} (published slug ${current.publishedSlug})${
-									current.live ? '' : ' ⚠️ not published — new workspaces fall back to the default'
-								}`
-							: 'none — using the built-in default'}
-				</span>
-			</div>
-			<div className={styles.searchContainer}>
-				<input
-					type="text"
-					placeholder="Published file ID"
-					ref={inputRef}
-					className={styles.searchInput}
-				/>
-				<TlaButton onClick={onSet} variant="primary">
-					Set as welcome template
-				</TlaButton>
-				<TlaButton onClick={onClear} variant="secondary" disabled={!current}>
-					Clear
-				</TlaButton>
-			</div>
-		</div>
-	)
-}
-
-function FeatureFlags() {
-	const [flags, setFlags] = useState<Record<string, FeatureFlagValue>>({})
-	const [isLoading, setIsLoading] = useState(true)
-	const [isSaving, setIsSaving] = useState(false)
-	const [error, setError] = useState(null as string | null)
-	const [successMessage, setSuccessMessage] = useState(null as string | null)
-
-	const loadFlags = useCallback(async () => {
-		setIsLoading(true)
-		setError(null)
-		try {
-			const res = await fetch('/api/app/admin/feature-flags')
-			if (!res.ok) {
-				setError(res.statusText + ': ' + (await res.text()))
-				return
-			}
-			const data = await res.json()
-			setFlags(data)
-		} catch (err) {
-			setError(err instanceof Error ? err.message : 'Failed to load flags')
-		} finally {
-			setIsLoading(false)
-		}
-	}, [])
-
-	useEffect(() => {
-		loadFlags()
-	}, [loadFlags])
-
-	const saveFlag = useCallback(
-		async (flag: string, update: { enabled?: boolean; percentage?: number }) => {
-			setIsSaving(true)
-			setError(null)
-			setSuccessMessage(null)
-			try {
-				const res = await fetch('/api/app/admin/feature-flags', {
-					method: 'POST',
-					headers: { 'Content-Type': 'application/json' },
-					body: JSON.stringify({ flag, ...update }),
-				})
-				if (!res.ok) {
-					setError(res.statusText + ': ' + (await res.text()))
-					return
-				}
-				setFlags((prev) => ({ ...prev, [flag]: { ...prev[flag], ...update } }))
-				if (update.percentage !== undefined) {
-					setSuccessMessage(
-						update.percentage === 0
-							? `${flag} disabled (0%)`
-							: `${flag} set to ${update.percentage}% of users`
-					)
-				} else {
-					setSuccessMessage(`${flag} ${update.enabled ? 'enabled' : 'disabled'}`)
-				}
-			} catch (err) {
-				setError(err instanceof Error ? err.message : 'Failed to update flag')
-			} finally {
-				setIsSaving(false)
-			}
-		},
-		[]
-	)
-
-	useEffect(() => {
-		if (successMessage) {
-			const timer = setTimeout(() => setSuccessMessage(null), 3000)
-			return () => clearTimeout(timer)
-		}
-	}, [successMessage])
-
-	return (
-		<div className={styles.fileOperation}>
-			{error && <div className={styles.errorMessage}>{error}</div>}
-			{successMessage && <div className={styles.successMessage}>{successMessage}</div>}
-
-			<p className={`tla-text_ui__small ${styles.featureFlagsNote}`}>
-				<strong>Global feature toggles.</strong> Changes take effect immediately for ALL users.
-			</p>
-			<p className={`tla-text_ui__small ${styles.featureFlagsDescription}`}>
-				Boolean flags toggle on/off for everyone. Percentage flags roll out to X% of users
-				(evaluated server-side per userId).
-			</p>
-
-			{isLoading ? (
-				<p className="tla-text_ui__small">Loading flags...</p>
-			) : (
-				<div className={styles.featureFlagsContainer}>
-					{Object.entries(flags)
-						.sort(([a], [b]) => {
-							// boolean flags first, then percentage flags
-							const aType = flags[a].type ?? 'boolean'
-							const bType = flags[b].type ?? 'boolean'
-							if (aType !== bType) return aType === 'boolean' ? -1 : 1
-							return a.localeCompare(b)
-						})
-						.map(([flagName, flagValue]) => {
-							const label = flagName
-								.split('_')
-								.map((word) => word.charAt(0).toUpperCase() + word.slice(1))
-								.join(' ')
-
-							if (flagValue.type === 'percentage') {
-								return (
-									<PercentageFlag
-										key={flagName}
-										flagName={flagName}
-										label={label}
-										flagValue={flagValue}
-										isSaving={isSaving}
-										onToggle={(enabled) => {
-											const action = enabled ? 'Enable' : 'Disable'
-											if (!window.confirm(`${action} "${flagName}"?`)) return
-											saveFlag(flagName, { enabled })
-										}}
-										onSavePercentage={(pct) => {
-											if (!window.confirm(`Set "${flagName}" to ${pct}% of users?`)) return
-											saveFlag(flagName, { percentage: pct })
-										}}
-									/>
-								)
-							}
-
-							return (
-								<div key={flagName} className={styles.featureFlagItem}>
-									<label htmlFor={flagName} className={styles.featureFlagLabel}>
-										<input
-											id={flagName}
-											type="checkbox"
-											checked={flagValue.enabled}
-											onChange={(e) => {
-												const enabled = e.target.checked
-												const action = enabled ? 'enable' : 'disable'
-												if (
-													!window.confirm(
-														`Are you sure you want to ${action} "${flagName}" for ALL users?`
-													)
-												) {
-													return
-												}
-												saveFlag(flagName, { enabled })
-											}}
-											disabled={isSaving}
-										/>
-										<span className="tla-text_ui__small">
-											<strong>{label}</strong>
-										</span>
-									</label>
-									{flagValue.description && (
-										<span className={`tla-text_ui__small ${styles.featureFlagsDescription}`}>
-											{flagValue.description}
-										</span>
-									)}
-								</div>
-							)
-						})}
-				</div>
-			)}
-		</div>
-	)
-}
-
-function PercentageFlag({
-	flagName,
-	label,
-	flagValue,
-	isSaving,
-	onToggle,
-	onSavePercentage,
-}: {
-	flagName: string
-	label: string
-	flagValue: PercentageFeatureFlag
-	isSaving: boolean
-	onToggle(enabled: boolean): void
-	onSavePercentage(percentage: number): void
-}) {
-	const currentPct = flagValue.percentage
-	const [pct, setPct] = useState(currentPct)
-
-	useEffect(() => {
-		setPct(currentPct)
-	}, [currentPct])
-
-	return (
-		<div className={styles.featureFlagItem}>
-			<div className={styles.featureFlagLabel}>
-				<label
-					htmlFor={flagName}
-					style={{ cursor: 'pointer', display: 'flex', alignItems: 'center', gap: 8 }}
-				>
-					<input
-						id={flagName}
-						type="checkbox"
-						checked={flagValue.enabled}
-						onChange={(e) => onToggle(e.target.checked)}
-						disabled={isSaving}
-						style={{ cursor: 'pointer' }}
-					/>
-					<span className="tla-text_ui__small">
-						<strong>{label}</strong>
-					</span>
-				</label>
-				<input
-					type="text"
-					value={pct}
-					onChange={(e) => {
-						const n = Number(e.target.value)
-						if (!Number.isNaN(n)) setPct(Math.max(0, Math.min(100, n)))
-					}}
-					disabled={isSaving || !flagValue.enabled}
-					className={styles.searchInput}
-					style={{ width: 60 }}
-				/>
-				<span
-					className={`tla-text_ui__small ${!flagValue.enabled ? styles.featureFlagDisabled : ''}`}
-				>
-					%
-				</span>
-				<TlaButton
-					onClick={() => onSavePercentage(pct)}
-					variant="primary"
-					disabled={isSaving || !flagValue.enabled || pct === currentPct}
-				>
-					Save
-				</TlaButton>
-			</div>
-			{flagValue.description && (
-				<span className={`tla-text_ui__small ${styles.featureFlagsDescription}`}>
-					{flagValue.description}
-				</span>
-			)}
-		</div>
-	)
-}
-
-function HardDeleteFile() {
-	const inputRef = useRef<HTMLInputElement>(null)
-	const [error, setError] = useState(null as string | null)
-	const [successMessage, setSuccessMessage] = useState(null as string | null)
-
-	const onDelete = useCallback(async () => {
-		const fileId = inputRef.current?.value
-		if (!fileId) {
-			setError('Please enter a file ID')
-			return
-		}
-
-		if (
-			!window.confirm(
-				`Are you sure you want to permanently delete file ${fileId}? This action cannot be undone.`
-			)
-		) {
-			return
-		}
-
-		setError(null)
-		setSuccessMessage(null)
-		const res = await fetch(`/api/app/admin/hard_delete_file/${fileId}`, {
-			method: 'POST',
-		})
-		if (!res.ok) {
-			setError(res.statusText + ': ' + (await res.text()))
-			return
-		} else {
-			setSuccessMessage('File deleted successfully! 🧹')
-			inputRef.current!.value = ''
-		}
-	}, [])
-
-	// Clear success message after 3 seconds
-	useEffect(() => {
-		if (successMessage) {
-			const timer = setTimeout(() => setSuccessMessage(null), 3000)
-			return () => clearTimeout(timer)
-		}
-	}, [successMessage])
-
-	return (
-		<div className={styles.dangerZone}>
-			{error && <div className={styles.errorMessage}>{error}</div>}
-			{successMessage && <div className={styles.successMessage}>{successMessage}</div>}
-			<div className={styles.deleteContainer}>
-				<input type="text" placeholder="File ID" ref={inputRef} className={styles.searchInput} />
-				<TlaButton onClick={onDelete} className={styles.deleteButton}>
-					Delete (cannot be undone)
-				</TlaButton>
-			</div>
-		</div>
-	)
-}
-
-function CreateLegacyFile() {
-	const [isCreating, setIsCreating] = useState(false)
-	const [successMessage, setSuccessMessage] = useState(null as string | null)
-
-	const handleCreate = useCallback(async () => {
-		setIsCreating(true)
-		setSuccessMessage(null)
-		try {
-			const res = await fetch(`/api/app/admin/create_legacy_file`, { method: 'POST' })
-			const { slug } = await res.json()
-			window.open(`/r/${slug}`, '_blank')?.focus()
-		} catch (err) {
-			console.error('Failed to create legacy file:', err)
-		} finally {
-			setIsCreating(false)
-		}
-	}, [])
-
-	// Clear success message after 3 seconds
-	useEffect(() => {
-		if (successMessage) {
-			const timer = setTimeout(() => setSuccessMessage(null), 3000)
-			return () => clearTimeout(timer)
-		}
-	}, [successMessage])
-
-	return (
-		<div className={styles.fileOperation}>
-			{successMessage && <div className={styles.successMessage}>{successMessage}</div>}
-			<TlaButton onClick={handleCreate} variant="secondary" isLoading={isCreating}>
-				Create Legacy File
-			</TlaButton>
-		</div>
-	)
-}
-
-function DownloadTldrFile({ legacy }: { legacy: boolean }) {
-	const inputRef = useRef<HTMLInputElement>(null)
-	const [error, setError] = useState(null as string | null)
-	const [isDownloading, setIsDownloading] = useState(false)
-	const [successMessage, setSuccessMessage] = useState(null as string | null)
-
-	const onDownload = useCallback(async () => {
-		setError(null)
-		setSuccessMessage(null)
-		const fileSlug = inputRef.current?.value
-		if (!fileSlug) {
-			setError('Please enter a file slug')
-			return
-		}
-		const path = legacy ? 'download-legacy-tldr' : 'download-tldr'
-
-		setIsDownloading(true)
-		try {
-			const res = await fetch(`/api/app/admin/${path}/${fileSlug}`)
-			if (!res.ok) {
-				setError(res.statusText + ': ' + (await res.text()))
-				return
-			}
-
-			// Create a blob from the response and trigger download
-			const blob = await res.blob()
-			const url = window.URL.createObjectURL(blob)
-			const a = document.createElement('a')
-			a.href = url
-			a.download = `${fileSlug}.tldr`
-			document.body.appendChild(a)
-			a.click()
-			window.URL.revokeObjectURL(url)
-			document.body.removeChild(a)
-		} finally {
-			setIsDownloading(false)
-		}
-	}, [legacy])
-
-	// Clear success message after 3 seconds
-	useEffect(() => {
-		if (successMessage) {
-			const timer = setTimeout(() => setSuccessMessage(null), 3000)
-			return () => clearTimeout(timer)
-		}
-	}, [successMessage])
-
-	return (
-		<div className={styles.fileOperation}>
-			<h4 className="tla-text_ui__medium">
-				{legacy ? 'Download Legacy .tldr File' : 'Download .tldr File'}
-			</h4>
-			{error && <div className={styles.errorMessage}>{error}</div>}
-			{successMessage && <div className={styles.successMessage}>{successMessage}</div>}
-			<div className={styles.downloadContainer}>
-				<input type="text" placeholder="File ID" ref={inputRef} className={styles.searchInput} />
-				<TlaButton onClick={onDownload} variant="primary" isLoading={isDownloading}>
-					Download
-				</TlaButton>
-			</div>
-		</div>
-	)
-}
-
-function formatBytes(bytes: number) {
-	if (bytes < 1024) return `${bytes} B`
-	if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`
-	return `${(bytes / (1024 * 1024)).toFixed(1)} MB`
-}
-
-function AssetDiagnostics() {
-	const inputRef = useRef<HTMLInputElement>(null)
-	const [error, setError] = useState(null as string | null)
-	const [isLoading, setIsLoading] = useState(false)
-	const [report, setReport] = useState(null as AdminFileAssetsResponseBody | null)
-
-	const onCheck = useCallback(async () => {
-		const slug = inputRef.current?.value?.trim()
-		if (!slug) {
-			setError('Please enter a file slug')
-			return
-		}
-		setError(null)
-		setReport(null)
-		setIsLoading(true)
-		try {
-			const res = await fetch(`/api/app/admin/file-assets/${encodeURIComponent(slug)}`)
-			if (!res.ok) {
-				setError(res.statusText + ': ' + (await res.text()))
-				return
-			}
-			setReport((await res.json()) as AdminFileAssetsResponseBody)
-		} catch (err) {
-			setError(err instanceof Error ? err.message : 'Failed to check assets')
-		} finally {
-			setIsLoading(false)
-		}
-	}, [])
-
-	return (
-		<div className={styles.fileOperation}>
-			<h4 className="tla-text_ui__medium">Asset diagnostics</h4>
-			<p className="tla-text_ui__regular">
-				Checks whether each asset in the file&apos;s last persisted snapshot exists in the uploads
-				bucket and is associated with the file.
-			</p>
-			{error && <div className={styles.errorMessage}>{error}</div>}
-			<div className={styles.searchContainer}>
-				<input
-					type="text"
-					placeholder="File slug"
-					ref={inputRef}
-					className={styles.searchInput}
-					onKeyDown={(e) => {
-						if (e.key === 'Enter') onCheck()
-					}}
-				/>
-				<TlaButton onClick={onCheck} variant="primary" isLoading={isLoading}>
-					Check assets
-				</TlaButton>
-			</div>
-			{report && (
-				<>
-					{report.warnings.length > 0 && (
-						<div className={styles.errorMessage}>
-							{report.warnings.length} check(s) failed — counts below may be incomplete.{' '}
-							{report.warnings.slice(0, 3).join('; ')}
-						</div>
-					)}
-					<div className={styles.userSummary}>
-						<div className={styles.summaryGrid}>
-							{[
-								[
-									'Shapes',
-									`${report.shapes.total}${
-										report.shapes.total > 0
-											? ` (${Object.entries(report.shapes.byType)
-													.sort((a, b) => b[1] - a[1])
-													.map(([type, count]) => `${count} ${type}`)
-													.join(', ')})`
-											: ''
-									}`,
-								],
-								['Total assets', report.assets.total],
-								[
-									'Asset size',
-									`${formatBytes(report.assets.totalSizeBytes)} (largest ${formatBytes(report.assets.largestSizeBytes)})`,
-								],
-								['Associated', report.assets.associated],
-								['Pending association', report.assets.pending],
-								['External (bookmarks etc.)', report.assets.external],
-								['Missing in bucket', report.assets.missingInBucket],
-								['Head check failures', report.assets.headFailures],
-								['Old-format URLs', report.assets.oldFormatUrls],
-								[
-									'DB asset rows',
-									`${report.dbRows.forThisFile} (${report.dbRows.orphaned} orphaned)`,
-								],
-								[
-									'Create source',
-									report.source
-										? `${report.source.raw} ${
-												report.source.exists === null
-													? '(not checked)'
-													: report.source.exists
-														? '(exists)'
-														: '⚠️ (missing)'
-											}`
-										: 'none',
-								],
-							].map(([label, value]) => (
-								<div key={label} className={styles.summaryItem}>
-									<span className={styles.fieldLabel}>{label}:</span>
-									<span className={styles.fieldValue}>{value}</span>
-								</div>
-							))}
-						</div>
-					</div>
-					{report.assets.problems.length > 0 && (
-						<table className={styles.diagnosticsTable}>
-							<thead>
-								<tr>
-									<th>Asset</th>
-									<th>Object name</th>
-									<th>In bucket</th>
-									<th>Meta fileId</th>
-									<th>DB fileId</th>
-								</tr>
-							</thead>
-							<tbody>
-								{report.assets.problems.map((p) => (
-									<tr key={p.assetId}>
-										<td>{p.assetId}</td>
-										<td>{p.objectName}</td>
-										<td>{p.inBucket === null ? 'check failed' : p.inBucket ? 'yes' : 'MISSING'}</td>
-										<td>{p.fileIdMeta ?? 'none'}</td>
-										<td>{p.dbRow?.fileId ?? 'none'}</td>
-									</tr>
-								))}
-							</tbody>
-						</table>
-					)}
-					<StructuredDataDisplay data={report} />
-				</>
-			)}
-		</div>
-	)
-}
-
-function DeleteUser() {
-	const inputRef = useRef<HTMLInputElement>(null)
-	const [isDeleting, setIsDeleting] = useState(false)
-	const [progressLog, setProgressLog] = useState<string[]>([])
-	const [error, setError] = useState(null as string | null)
-	const [isComplete, setIsComplete] = useState(false)
-
-	const onDelete = useCallback(async () => {
-		const userId = inputRef.current?.value?.trim()
-		if (!userId) {
-			setError('Please enter a user ID or email')
-			return
-		}
-
-		if (
-			!window.confirm(
-				`Are you sure you want to permanently delete user "${userId}"? This action cannot be undone and will delete all their files, data, and account.`
-			)
-		) {
-			return
-		}
-
-		setIsDeleting(true)
-		setError(null)
-		setProgressLog([]) // Only clear log when starting a new deletion
-		setIsComplete(false)
-
-		try {
-			const eventSource = new EventSource(
-				`/api/app/admin/delete_user_sse?q=${encodeURIComponent(userId)}`
-			)
-
-			eventSource.onmessage = (event) => {
-				const data = JSON.parse(event.data)
-
-				const timestamp = new Date(data.timestamp).toLocaleTimeString()
-				const logEntry = `[${timestamp}] ${data.message}`
-
-				setProgressLog((prev) => [...prev, logEntry])
-
-				if (data.type === 'complete') {
-					setIsComplete(true)
-					setIsDeleting(false)
-					eventSource.close()
-				} else if (data.type === 'error') {
-					setError(data.message)
-					setIsDeleting(false)
-					eventSource.close()
-				}
-			}
-
-			eventSource.onerror = () => {
-				setError('Connection failed')
-				setIsDeleting(false)
-				eventSource.close()
-			}
-		} catch (err) {
-			setError(err instanceof Error ? err.message : 'Unknown error occurred')
-			setIsDeleting(false)
-		}
-	}, [])
-
-	return (
-		<div className={styles.dangerZone}>
-			<h4>Delete User</h4>
-			{error && <div className={styles.errorMessage}>{error}</div>}
-			{isComplete && <div className={styles.successMessage}>User deleted successfully! 🧹</div>}
-
-			<div className={styles.deleteContainer}>
-				<input
-					type="text"
-					placeholder="User ID or Email"
-					ref={inputRef}
-					className={styles.searchInput}
-					disabled={isDeleting}
-				/>
-				<TlaButton
-					onClick={onDelete}
-					className={styles.deleteButton}
-					disabled={isDeleting}
-					isLoading={isDeleting}
-				>
-					{isDeleting ? 'Deleting...' : 'Delete User (cannot be undone)'}
-				</TlaButton>
-			</div>
-
-			{/* Progress Log */}
-			{progressLog.length > 0 && (
-				<div className={styles.progressLog}>
-					<h5>Deletion Progress:</h5>
-					<div className={styles.logContainer}>
-						{progressLog.map((log, index) => (
-							<div key={index} className={styles.logEntry}>
-								{log}
-							</div>
-						))}
-					</div>
-				</div>
-			)}
-		</div>
-	)
-}
